requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.0
django-admin-search-builder==0.2.2
django-allauth==0.50.0
django-cors-headers==3.11.0
django-debug-panel==0.8.3
django-debug-toolbar==3.2.4
django-ipware==4.0.2
django-rest-swagger==2.2.0
django-silk==4.3.0
djangorestframework==3.13.1
drf-yasg==1.20.0
envparse==0.2.0
et-xmlfile==1.1.0
gprof2dot==2021.2.21
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
MarkupSafe==2.1.1
numpy
oauthlib==3.2.0
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
pandas
psycopg2-binary==2.8.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.9.1
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.10
ipython==7.16.1
